+++ frontend/src/components/onboard/onboard.tsx
@@ -7,6 +7,7 @@ import { Colors } from '../../config/colors'
 import { DEFAULT_VEGA_HOME } from '../../config/environment'
 import { Intent } from '../../config/intent'
 import { useGlobal } from '../../contexts/global/global-context'
+import { useNetwork } from '../../contexts/network/network-context'
 import { useCreateWallet } from '../../hooks/use-create-wallet'
 import { useImportWallet } from '../../hooks/use-import-wallet'
 import { OnboardPaths, Paths } from '../../routes'
@@ -29,25 +30,79 @@ export function Onboard() {
 
 export function OnboardHome() {
   const navigate = useNavigate()
-  const [loading, setLoading] = React.useState<'create' | 'import' | null>(null)
+  const [loading, setLoading] = React.useState<
+    'create' | 'import' | 'existing' | null
+  >(null)
   const {
-    state: { version }
+    dispatch: globalDispatch,
+    state: { version, onboarding }
   } = useGlobal()
+  const { dispatch: networkDispatch } = useNetwork()
+
+  const initialiseWithDefaultHome = async () => {
+    try {
+      await Service.InitialiseApp({ vegaHome: DEFAULT_VEGA_HOME })
+    } catch (err) {
+      Sentry.captureException(err)
+    }
+  }
+
+  const handleImportExistingWallet = async () => {
+    try {
+      setLoading('existing')
+
+      await Service.InitialiseApp({ vegaHome: DEFAULT_VEGA_HOME })
+
+      const [wallets, networks] = await Promise.all([
+        Service.ListWallets(),
+        Service.ListNetworks()
+      ])
+
+      globalDispatch({ type: 'ADD_WALLETS', wallets: wallets.wallets })
+      networkDispatch({
+        type: 'ADD_NETWORKS',
+        networks: networks.networks
+      })
+      globalDispatch({ type: 'INIT_APP', isInit: true })
+
+      // If use doesnt have networks go to the import network section on onboarding
+      // otherwise go to home to complete onboarding
+      if (!onboarding.networks.length) {
+        navigate(OnboardPaths.Network)
+      } else {
+        navigate(Paths.Home)
+      }
+    } catch (err) {
+      Sentry.captureException(err)
+    }
+  }
 
   return (
     <div style={{ textAlign: 'center' }}>
       <Header style={{ margin: '0 0 30px 0', color: Colors.WHITE }}>
         <Vega />
       </Header>
+      {onboarding.wallets.length ? (
+        <p style={{ maxWidth: 370 }}>
+          Existing wallet found. Either use existing, create a new wallet or
+          import from your recovery phrase
+        </p>
+      ) : null}
       <ButtonGroup orientation='vertical' style={{ marginBottom: 20 }}>
+        {onboarding.wallets.length || onboarding.networks.length ? (
+          <Button
+            loading={loading === 'existing'}
+            onClick={handleImportExistingWallet}
+          >
+            Use existing
+          </Button>
+        ) : null}
         <Button
           loading={loading === 'create'}
           data-testid='create-new-wallet'
           onClick={async () => {
             setLoading('create')
-            await Service.InitialiseApp({
-              vegaHome: DEFAULT_VEGA_HOME
-            })
+            await initialiseWithDefaultHome()
             navigate(OnboardPaths.WalletCreate)
           }}
         >
@@ -58,9 +113,7 @@ export function OnboardHome() {
           loading={loading === 'import'}
           onClick={async () => {
             setLoading('import')
-            await Service.InitialiseApp({
-              vegaHome: DEFAULT_VEGA_HOME
-            })
+            await initialiseWithDefaultHome()
             navigate(OnboardPaths.WalletImport)
           }}
         >
@@ -97,10 +150,9 @@ export function OnboardSettings() {
           vegaHome: values.vegaHome
         })
         AppToaster.show({ message: 'App initialised', intent: Intent.SUCCESS })
-        navigate('/')
+        navigate(Paths.Onboard)
       } catch (err) {
         Sentry.captureException(err)
-        console.error(err)
         setLoading(false)
       }
     },
@@ -108,7 +160,7 @@ export function OnboardSettings() {
   )
 
   return (
-    <OnboardPanel title='Advanced settings' back={OnboardPaths.Home}>
+    <OnboardPanel title='Advanced settings' back={Paths.Onboard}>
       <form onSubmit={handleSubmit(submit)}>
         <FormGroup
           label='Vega home'
@@ -117,11 +169,12 @@ export function OnboardSettings() {
         >
           <input type='text' {...register('vegaHome')} />
         </FormGroup>
-        <div>
+        <ButtonGroup>
           <Button type='submit' loading={loading}>
             Initialise
           </Button>
-        </div>
+          <Button onClick={() => navigate(-1)}>Cancel</Button>
+        </ButtonGroup>
       </form>
     </OnboardPanel>
   )
@@ -132,7 +185,7 @@ export function OnboardWalletCreate() {
   const { submit, response } = useCreateWallet()
 
   return (
-    <OnboardPanel title='Create wallet' back={OnboardPaths.Home}>
+    <OnboardPanel title='Create wallet' back={Paths.Onboard}>
       {response ? (
         <WalletCreateFormSuccess
           response={response}
@@ -146,10 +199,7 @@ export function OnboardWalletCreate() {
           }
         />
       ) : (
-        <WalletCreateForm
-          submit={submit}
-          cancel={() => navigate(OnboardPaths.Home)}
-        />
+        <WalletCreateForm submit={submit} cancel={() => navigate(-1)} />
       )}
     </OnboardPanel>
   )
@@ -166,11 +216,8 @@ export function OnboardWalletImport() {
   }, [response, navigate])
 
   return (
-    <OnboardPanel title='Import a wallet' back={OnboardPaths.Home}>
-      <WalletImportForm
-        submit={submit}
-        cancel={() => navigate(OnboardPaths.Home)}
-      />
+    <OnboardPanel title='Import a wallet' back={Paths.Onboard}>
+      <WalletImportForm submit={submit} cancel={() => navigate(-1)} />
     </OnboardPanel>
   )
 }
@@ -180,7 +227,7 @@ export function OnboardNetwork() {
   const { dispatch } = useGlobal()
 
   const onComplete = React.useCallback(() => {
-    dispatch({ type: 'FINISH_ONBOARDING' })
+    dispatch({ type: 'INIT_APP', isInit: true })
     navigate(Paths.Home)
   }, [dispatch, navigate])
 
@@ -194,7 +241,7 @@ export function OnboardNetwork() {
 interface OnboardPanelProps {
   children: React.ReactNode
   title: React.ReactNode
-  back: OnboardPaths
+  back: string
 }
 
 export function OnboardPanel({ children, title, back }: OnboardPanelProps) {